refactored ConfigStore to match camel case
