+++ src/plugins/plot/configuration/ConfigStore.js
